Added prediction to labels
