+++ Content.Client/Labels/EntitySystems/LabelSystem.cs
@@ -1,0 +1,7 @@
+using Content.Shared.Labels.EntitySystems;
+
+namespace Content.Client.Labels;
+
+public sealed partial class LabelSystem : SharedLabelSystem
+{
+}
